+++ CHANGELOG.md
@@ -2,11 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.3.0](https://github.com/julie-ng/nodejs-client-certificate-auth/compare/v0.2.0...v0.3.0) (2020-03-06)
+
+
+### Features
+
+* **azure:** add example using key vault certificates, as opposed to secrets ([74ec2e6](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/74ec2e6f84279ddbdd9b2a5d3ebc32364164e094))
+* **server:** add message indicating startup ([ac7ec66](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/ac7ec66a9fa66cb79dc689b67ee13862fac3ea92))
+
+## [0.2.0](https://github.com/julie-ng/nodejs-client-certificate-auth/compare/v0.1.0...v0.2.0) (2020-03-05)
+
+
+### Features
+
+* **azure:** refactor, add sample .env, docs ([708c2fc](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/708c2fc5a0ff41361824c7799e8ce1e332519539))
+* **azure:** use key vault secrets for certificates ([99ff1fe](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/99ff1feefbdb9b6ccd754d9d292ed6f8ac06bc27))
+* **certs:** add script to generate ([f0dae2a](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/f0dae2a355f3cef78c1e4f7de125543db2d67bce))
+
 ### 0.1.0 (2018-09-19)
 
 _Note: this release was retroactively tagged_
 
 ### Features
 
-* **client**: add invalida dn valid demo apps ([d8dda3f](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/d8dda3fd9ada4494e5bc85d9717745eb7c9fa292), [b93708e](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/b93708e9b897e5f1ee15738610a1284b240e8b30))
+* **client**: add invalid and valid demo apps ([d8dda3f](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/d8dda3fd9ada4494e5bc85d9717745eb7c9fa292), [b93708e](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/b93708e9b897e5f1ee15738610a1284b240e8b30))
 * **server**: add ([0b61c77](https://github.com/julie-ng/nodejs-client-certificate-auth/commit/0b61c77b208571ddb625975c2ef20bbc83640bb0))